backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0

# GitHub
PyGithub==2.5.0
PyJWT[crypto]==2.10.1

# LLM Providers
groq==0.15.0
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
structlog==24.4.0
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
